+++ netket/vqs/mc/mc_mixed_state/__init__.py
@@ -12,14 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .base import VariationalState, VariationalMixedState, expect, expect_and_grad
-
-from .mc import MCState, MCMixedState
-
-# TODO: this is deprecated in favour of netket.experimental.vqs
-# eventually remove this file and import
-from . import experimental
-
-from netket.utils import _hide_submodules
-
-_hide_submodules(__name__, ignore=["experimental"])
+from .state import MCMixedState
